feat: config.remove and config.add declaration
Using a signal was the way to go, really proud of myself for this code
:)
The implementation will happen as soon as I get the abstract aspect out
of the way.
The key-file implementation was already kind of there, but I would like
to understand threading/loops better to implement a file monitor too.